requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=4.0.0
pyjwt>=2.6.0
drf-yasg>=1.21.5
gunicorn>=21.2.0
whitenoise>=6.4.0
python-decouple>=3.8
django-heroku>=0.3.1


